fix(tools): do not include resource in the tools payload (#11)
Including the full resources in the tools payload results in an 
extremely large resource size, which causes the mcp server
response to time out (e.g. with this example from claude-desktop):

```
2025-05-12T05:04:56.699Z [bookstore] [info] Message from client: {"jsonrpc":"2.0","method":"notifications/cancelled","params":{"requestId":1,"reason":"Error: MCP error -32001: Request timed out"}}
```

The fix is to remove the resource from the payload, and map the tool
to the request another way.